Add a couple more unit tests to the Mac build Review URL: http://codereview.chromium.org/43076
git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11446 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/browser_commands_unittest.cc
@@ -59,7 +59,7 @@ TEST_F(BrowserCommandsTest, DuplicateTab) {
   browser()->ExecuteCommand(IDC_DUPLICATE_TAB);
 
   // The duplicated tab should not end up in a new window.
-  int window_count = BrowserList::size();
+  size_t window_count = BrowserList::size();
   ASSERT_EQ(initial_window_count, window_count);
 
   // And we should have a newly duplicated tab.
